requirements.txt
attrs==23.1.0
cattrs==23.2.3
certifi==2023.11.17
charset-normalizer==3.3.2
exceptiongroup==1.2.0
flatbuffers==23.5.26
idna==3.6
iniconfig==2.0.0
numpy==1.26.2
openmeteo_requests==1.1.0
openmeteo_sdk==1.7.0
packaging==23.2
pandas==2.1.4
platformdirs==4.1.0
pluggy==1.3.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
requests-cache==1.1.1
retry-requests==2.0.0
six==1.16.0
suncalc==0.1.3
tomli==2.0.1
typing_extensions==4.9.0
tzdata==2023.3
url-normalize==1.4.3
urllib3==2.1.0
